#137 when all bans are done, new player order is generated for picks
